[metadata] licenses: add `license_kind` rule and provider (#13)
This change introduces `LicenseKindInfo` and `license_kind`, which will
be used to uniquely identify a license.

This differs from `@rules_licence`'s `license_kind` in that it only
provides a the identifier and (human readable) name for a license kind
and removes all other attributes (most notably `conditions`, which is
always specific to an organization).

Splitting the identifier of a license from its conditions has the
benefit that organizations can reuse the `license_kind` from the OSS
package and don't have to maintain patches for injecting their
conditions.

A follow-up change will introduce an (optional) toolchain for injecting
conditions which organizations can define.

```starlark
license_conditions_toolchain(
    name = "toolchain",
    conditions = [
        ":mit",
        ":lgpl",
        # ...
    ],
)

license_condition(
    name = "mit",
    license_kind = "@package_metadata//licenses/spdx:mit",
    conditions = [
        "notice",
    ],
)

license_condition(
    name = "lgpl",
    license_kind = "@package_metadata//licenses/spdx:lgpl",
    conditions = [
        "notice_when_dynamicly_linked",
    ],
)
```