Pull list of PacakgeAttributeInfo from children into PackageMetadataInfo.
This allows an aspect descending from a target to see the PackageAttributeInfo
instances for a `package_metadata` instance when it is attached to that target
via the common attribute `package_metadata`.

Review question: Should we drop the `files` field from PMI. It is no
redundant w.r.t. the `attributes` field.