Initial version of tools to gather pacakge metadata.
This is based on the license gatherer from rules_license.

It is not ready for use yet because it is not as efficient or useful
as it needs to be

- it is hooked into docs generation so that it may be easier to see what is going on.
- it updates some bits of the metadata license rule to use package_metadata rather
  than applicable_licenses. That has worked in all of Bazel 7.

This change only gets enough in place to emit a list of all the kinds
of attribute gathered and print paths of where the output was emitted.
But it gives us a place to try out new things from and to start to build
unit tests.

Next parts:
- Add the data that lets us reconsruct the dep graph. This is needed for nicer SBOMS

This is FUN!!! Consider this

```
package(default_package_metadata=["//:package_metadata"])

package_metadata(
   name = "package_metadata",
   attributes = [":license"],
   ...
)

cc_binary(name = "main", ...)

check_binary(name = "check", deps = [":main"])
```

Where check_binary is a rule that runs the aspect down the graph to collect
the PackageAttributeInfo things. I don't see evidence that I visit `:license`.
There is a lot of debug scaffolding in there, and it is baffling me.

Yannic and I were at odds about this from months ago. This PR provides enough
of an example to be a testbed for trying out ideas.  The basic problem is
that you can't act on any of the values of a PackageAttributeInfo in a rule.
You need to gather the file handles, then pass them to an action that does
what you need. That raises the bar a bit.

revert applicable_licenses deprecation

beep

try the new bit

Initial version of tools to gather pacakge metadata.

This is based on the license gatherer from rules_license.

It is not ready for use yet because it is not as efficient or useful
as it needs to be

- it is hooked into docs generation so that it may be easier to see what is going on.
- it updates some bits of the metadata license rule to use package_metadata rather
  than applicable_licenses. That has worked in all of Bazel 7.

This change only gets enough in place to emit a list of all the kinds
of attribute gathered and print paths of where the output was emitted.
But it gives us a place to try out new things from and to start to build
unit tests.

Next parts:
- Add the data that lets us reconsruct the dep graph. This is needed for nicer SBOMS

This is FUN!!! Consider this

```
package(default_package_metadata=["//:package_metadata"])

package_metadata(
   name = "package_metadata",
   attributes = [":license"],
   ...
)

cc_binary(name = "main", ...)

check_binary(name = "check", deps = [":main"])
```

Where check_binary is a rule that runs the aspect down the graph to collect
the PackageAttributeInfo things. I don't see evidence that I visit `:license`.
There is a lot of debug scaffolding in there, and it is baffling me.

Yannic and I were at odds about this from months ago. This PR provides enough
of an example to be a testbed for trying out ideas.  The basic problem is
that you can't act on any of the values of a PackageAttributeInfo in a rule.
You need to gather the file handles, then pass them to an action that does
what you need. That raises the bar a bit.

revert applicable_licenses deprecation

beep

try the new bit