Add kind and purl to PackageMetadataInfo
kind helps when dispatching on gathered metadata when mixing out stuff
with company private metadata. It is a valid distinguisher to know
to pass that to sbom generation.

Having the purl in the provide helps with debugging and demos,
and simply being able to produce useful output during analysis time.