Fix broken package metadata collector in core.bzl (#64)
Fix broken package metadata collector in core.bzl

- it now can collect mixed metadata and user defined providers
- they must obey a protocol
  - must include 'kind'
- if they include 'metadata' or 'attributes' those are considered paths
of JSON files. We pass them to the sbom bulider
- 'files' attributes are depsets of files to be passed to
ctx.actions.run.
- Make it more efficient
- Improve the example to help prove out this code.

Add kind and purl to package_metadata

There is more to do. I don't like the design. The single collection
provider should be two. One which is just the depset. THe other which is
interesting targets.