Update spdx license list (#86)
Run the script from rules_license to bring us up to date with the
published list on spdx.org

Doing this for temporary completeness, but the destiny is not clear.
It should rev differently than metadata, or tools. Perhaps a separate
repo just for that because we can auto-release it easily.