[metadata] Implement encoder for purls (part 1) (#38)
This will make it easier for users (esp. bzlmod extensions integrating
with package managers) to generate package_metadata targets.

The new method (`string.to_bytes()`) must be used on distinct components
of a PURL, and not on the entire url.
Read
https://github.com/package-url/purl-spec/blob/main/purl-specification.md
carefully before using it.

Based on https://github.com/bazel-contrib/supply-chain/pull/14

---------

Co-authored-by: Yannic Bonenberger <yannic@engflow.com>
Co-authored-by: Yannic Bonenberger <contact@yannic-bonenberger.com>